src/webui/pyproject.toml
[tool.poetry]
name = "webui"
version = "0.1.0"
description = ""
authors = ["rcastillo <tepuybits@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
jinja2 = "^3.1.3"
nats-py = "^2.7.2"
websockets = "^12.0"
pandas = "^2.2.1"
uvicorn = "^0.29.0"
fastapi = "^0.110.0"
python-arango = "^7.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
